feat: add workflows to check conformity with prettier and eslint configs (#46)
These new workflows will run `prettier` and `eslint` on code when a PR
is created, to make
sure that messy code doesn't make it into the repo. This is another step
of verification
on top of the commit hooks. These workflows only check compliance, they
do not actually
change the code. To resolve failures, one must make another commit to
the PR to fix any
issues.

---------

Co-authored-by: Dylan Tompkins <dltompki@outlook.com>